backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5
langchain==0.0.335
chromadb==0.4.18
sentence-transformers==2.2.2
html2text==2020.1.16
ollama==0.1.4
huggingface-hub==0.19.4
tiktoken==0.5.1
python-multipart>=0.0.5
python-dotenv==1.0.0 